requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlite3==3.42.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6